src/hanson/requirements.txt
wheel
Mastodon.py
discord
PyGithub
python-dotenv
pygit2
#pygraphviz
python_terraform